requirements.txt
google-cloud-pubsub==2.7.0
psycopg2==2.9.1
python-dotenv==0.19.1